Implement interface for espay client
