+++ espay/signature.go
@@ -1,0 +1,5 @@
+package espay
+
+type ISignature interface {
+	CreateSignature(signatureKey string) string
+}
